Feedback 25606
-Removed Query component, replaced with useQuery usage.
-Updated docs.
