+++ packages/dina-ui/pages/object-store/metadata/single-record-edit.tsx
@@ -5,12 +5,12 @@ import {
   DinaForm,
   DinaFormOnSubmit,
   FieldSet,
-  Query,
   ResourceSelectField,
   SelectField,
   SubmitButton,
   TextField,
   useApiClient,
+  useQuery,
   withResponse
 } from "common-ui";
 import { Field } from "formik";
@@ -21,8 +21,8 @@ import {
   Head,
   Nav,
   NotPubliclyReleasableWarning,
-  TagsAndRestrictionsSection,
-  PersonSelectField
+  PersonSelectField,
+  TagsAndRestrictionsSection
 } from "../../../components";
 import { ManagedAttributesEditor } from "../../../components/object-store/managed-attributes/ManagedAttributesEditor";
 import { MetadataFileView } from "../../../components/object-store/metadata/MetadataFileView";
@@ -48,6 +48,35 @@ export default function MetadataEditPage() {
     query: { id }
   } = router;
 
+  const query = useQuery<Metadata>(
+    {
+      path: `objectstore-api/metadata/${id}`,
+      include: "dcCreator,derivatives"
+    },
+    {
+      joinSpecs: [
+        // Join to persons api:
+        {
+          apiBaseUrl: "/agent-api",
+          idField: "dcCreator",
+          joinField: "dcCreator",
+          path: metadata => `person/${metadata.dcCreator.id}`
+        }
+      ],
+      onSuccess: async ({ data: metadata }) => {
+        // Get the License resource based on the Metadata's xmpRightsWebStatement field:
+        if (metadata.xmpRightsWebStatement) {
+          const url = metadata.xmpRightsWebStatement;
+          (metadata as any).license = (
+            await apiClient.get<License[]>("objectstore-api/license", {
+              filter: { url }
+            })
+          ).data[0];
+        }
+      }
+    }
+  );
+
   return (
     <div>
       <Head title={formatMessage("editMetadataTitle")} />
@@ -58,41 +87,9 @@ export default function MetadataEditPage() {
             <h1 id="wb-cont">
               <DinaMessage id="editMetadataTitle" />
             </h1>
-            <Query<Metadata>
-              query={{
-                path: `objectstore-api/metadata/${id}`,
-                include: "dcCreator,derivatives"
-              }}
-              options={{
-                joinSpecs: [
-                  // Join to persons api:
-                  {
-                    apiBaseUrl: "/agent-api",
-                    idField: "dcCreator",
-                    joinField: "dcCreator",
-                    path: metadata => `person/${metadata.dcCreator.id}`
-                  }
-                ],
-                onSuccess: async ({ data: metadata }) => {
-                  // Get the License resource based on the Metadata's xmpRightsWebStatement field:
-                  if (metadata.xmpRightsWebStatement) {
-                    const url = metadata.xmpRightsWebStatement;
-                    (metadata as any).license = (
-                      await apiClient.get<License[]>(
-                        "objectstore-api/license",
-                        { filter: { url } }
-                      )
-                    ).data[0];
-                  }
-                }
-              }}
-            >
-              {metadataQuery =>
-                withResponse(metadataQuery, ({ data }) => (
-                  <SingleMetadataForm metadata={data} router={router} />
-                ))
-              }
-            </Query>
+            {withResponse(query, ({ data }) => (
+              <SingleMetadataForm metadata={data} router={router} />
+            ))}
           </div>
         )}
       </main>
